Move serverless to common parser node
